+++ lib/neo_pubspec.dart
@@ -1,6 +1,7 @@
 /// New generation of `pubspec.yaml` reader and writer
+///
+/// It allows modify pubspec and executing command on located directory.
 library neo_pubspec;
 
 export 'src/file.dart' hide ExecuteCmd;
 export 'src/structre/structre.dart' hide FlutterPubspecCondition;
-export 'src/validator.dart';