+++ lib/validator.dart
@@ -1,0 +1,7 @@
+/// A bundle of all validation for checking pubspec file
+///
+/// All rules of the validation is followed by
+/// [official documentation](https://dart.dev/tools/pub/pubspec).
+library neo_pubspec.validator;
+
+export 'src/validator.dart';